feat: display openings polygon on the map
jira issue silva-236
