+++ frontend/src/components/OpeningsMap/index.tsx
@@ -2,102 +2,39 @@ import React, { useEffect, useRef, useState } from 'react';
 import { useMap } from 'react-leaflet';
 import { Polygon } from 'react-leaflet';
 import { MapContainer, TileLayer, Marker, Popup, Rectangle } from 'react-leaflet';
+import { OpeningPolygon } from '../../types/OpeningPolygon';
+import { getOpeningsPolygonFromWfs } from '../../map-services/BcGwWfsApi';
 
 interface MapProps {
   selectedBasemap: any;
   openingId: number | null;
 }
 
-interface OpeningMap {
-  key: string;
-  bounds: any;
-  properties: object;
-  id: string;
-}
-
 const OpeningsMap: React.FC<MapProps> = ({ selectedBasemap, openingId }) => {
   //const position: [number, number] = [49.11941257871176, -122.83461402566606];
   const lastClickedLayerRef = useRef<any>(null); // Replace 'any' with the specific type if known
-  const [openings, setOpenings] = useState<OpeningMap[]>([]);
-  const [position, setPosition] = useState<number[]>([59.30407, -122.0567]);
+  const [openings, setOpenings] = useState<OpeningPolygon[]>([]);
+  const [position, setPosition] = useState<number[]>([49.2568449, -123.1289342]);
   const [reloadMap, setReloadMap] = useState<boolean>(false);
 
   const resultsStyle = {
     color: 'black'
   };
 
-  const invertXandY = (coordinates: any) => {
-    const newCoord = [];
-    for (let i = 0, len = coordinates.length; i < len; i++) {
-      const polygon = coordinates[i];
-      const newSubArray = []
-
-      for (let j = 0, lenj = polygon.length; j < lenj; j++) {
-        const lat = polygon[j][1];
-        const long = polygon[j][0];
-        const polygonPoint = [lat, long];
-        newSubArray.push(polygonPoint);
-      }
-      newCoord.push(newSubArray);
-    }
-    return newCoord;
-  }
-
-  const getOpeningsData = async (openingId: number) => {
-    let uri = 'https://openmaps.gov.bc.ca/geo/ows';
-    // service
-    uri += '?service=WFS';
-    // version
-    uri += '&version=2.0.0';
-    // request
-    uri += '&request=GetFeature';
-    // typeName (layer !?)
-    uri += '&typeName=WHSE_FOREST_VEGETATION.RSLT_OPENING_SVW';
-    // output format
-    uri += '&outputFormat=application/json';
-    // Srs name
-    uri += '&SrsName=EPSG:4326';
-    // Properties name
-    uri += '&PROPERTYNAME=OPENING_ID,APPROVE_DATE,LICENSEE_OPENING_ID,GEOMETRY';
-    // CQL Filters
-    uri += `&CQL_FILTER=OPENING_ID=${openingId}`;
-
-    const resultJson = await fetch(uri);
-    if (resultJson.ok) {
-      const json = await resultJson.json();
-
-      if (json.features && json.features.length) {
-        const openingsList: OpeningMap[] = [];
-        for (let i = 0, len = json.features.length; i < len; i++) {
-          if (json.features[i].geometry) {
-            // get the position from the bounding box
-            setPosition([json.bbox[1], json.bbox[0]]);
-
-            const openingGeometry = invertXandY(json.features[i].geometry.coordinates);
-
-            const openingObj: OpeningMap = {
-              key: json.features[i].id,
-              bounds: openingGeometry,
-              properties: json.features[i].properties,
-              id: json.features[i].id,
-            };
-            openingsList.push(openingObj);
-          }
-        }
-
-        if (openingsList.length) {
-          setOpenings(openingsList);
-          setReloadMap(true);
-        }
+  useEffect(() => {
+    const callBcGwApi = async () => {
+      const openingGeom: OpeningPolygon | null = await getOpeningsPolygonFromWfs(openingId);
+      if (openingGeom) {
+        setOpenings([openingGeom]);
+        setPosition([openingGeom.positionLat, openingGeom.positionLong]);
+        setReloadMap(true);
       } else {
         window.alert('No features found for the Opening ID ' + openingId);
       }
-    }
-  };
+    };
 
-  useEffect(() => {
     if (openingId) {
-      getOpeningsData(openingId);
+      callBcGwApi();
     }
   }, [openingId]);
 
@@ -116,7 +53,7 @@ const OpeningsMap: React.FC<MapProps> = ({ selectedBasemap, openingId }) => {
   return (
     <MapContainer
       center={position}
-      zoom={12}
+      zoom={13}
       style={{ height: "400px", width: "100%" }}
     >
       <TileLayer
