+++ frontend/src/types/OpeningPolygon.ts
@@ -1,0 +1,8 @@
+export type OpeningPolygon = {
+  key: string;
+  bounds: any;
+  properties: object;
+  id: string;
+  positionLat: number;
+  positionLong: number;
+};
